ajax + blueprint
* przeniesienie struktury my trashbox do katalogu include z prefixel include_
* includ podciąga ajax
* blueprinty są częścią app.py, znajdują się w katalogu blueprint
