+++ static/blueprints/testFile.py
@@ -1,0 +1,21 @@
+def getTestFiles():
+    paths = [{
+        'fileID': '1',
+        'Name': 'kociaki',
+        'Extension': '.png',
+        'FilePath': '/',
+        'Size': '240000',
+        'HashSum': '123',
+        'Icon': 'file-image',
+    },
+        {
+            'fileID': '2',
+            'Name': 'TrashPanda',
+            'Extension': 'None',
+            'FilePath': '/',
+            'Size': '2450000',
+            'HashSum': '4556',
+            'Icon': 'folder-open-empty',
+        },
+    ]
+    return paths
