--- static/scripts/js/auth.js
@@ -1,6 +1,0 @@
-if (localStorage.getItem('signed') != null) {
-    console.log("Login Active");
-} else {
-    window.localStorage.clear();
-    window.location.href = ("/");
-}
