+++ templates/extend/head.html
@@ -1,10 +1,25 @@
+{# google sign in #}
 <meta name="google-signin-client_id"
       content="557305630089-3sr6loe91a193q6t9e2du5np98rt0khf.apps.googleusercontent.com">
 <script src="https://apis.google.com/js/platform.js" async defer></script>
+
+{# ajax, jquery, itp #}
 <script src="//ajax.googleapis.com/ajax/libs/jquery/1.9.1/jquery.min.js"></script>
 <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
 <script src="https://ajax.aspnetcdn.com/ajax/jQuery/jquery-3.3.1.min.js"></script>
+
+{# skrypt pozwalający na używanie sesji z js'a #}
 <script src="/static/scripts/js/sessionFlask.js"></script>
+
+{# redirect script #}
+<script src="/static/scripts/js/goTo.js"></script>
+
+{# Ajaxowe podciąganie do div include #}
+<script async src="/static/scripts/js/ajaxChangePanel.js"></script>
+
+{# Kodowanie #}
 <meta charset="UTF-8">
+
+{# Czcionki #}
 <link href="https://fonts.googleapis.com/css?family=Aladin|Almendra+Display|Black+Ops+One|Cabin+Sketch:400,700|Caesar+Dressing|Creepster|East+Sea+Dokdo|Faster+One|Finger+Paint|Fredericka+the+Great|Gugi|IM+Fell+English+SC|Indie+Flower|Josefin+Slab:400,700|Kirang+Haerang|Libre+Barcode+39|Libre+Barcode+39+Extended+Text|Londrina+Sketch|Love+Ya+Like+A+Sister|Megrim|Merienda:400,700|Permanent+Marker|Press+Start+2P|Quantico:400,700|Rock+Salt|Shadows+Into+Light|Special+Elite|Trade+Winds|VT323|Yatra+One&amp;subset=latin-ext"
       rel="stylesheet">
